Rename Test1884 to Test1902 to fit with PR number
