+++ TestsExample/App.js
@@ -92,7 +92,7 @@ import Test1791 from './src/Test1791';
 import Test1802 from './src/Test1802';
 import Test1844 from './src/Test1844';
 import Test1864 from './src/Test1864';
-import Test1884 from './src/Test1884';
+import Test1902 from './src/Test1902';
 
 enableFreeze(true);
